requirements.txt
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pillow==8.2.0
Werkzeug==2.0.2
numpy>=1.9.2
WSGIserver==1.3
psycopg2==2.9.1
torchvision==0.11.1